Car Details Page

+++ server/database/config/build.sql
@@ -41,7 +41,7 @@ CREATE TABLE rentals (
     location TEXT,
     pick_up_date_time TIMESTAMPTZ,
     pick_off_date_time TIMESTAMPTZ,
-    total_price INT
+    total_price DECIMAL
 );
 CREATE TABLE reviews (
     id SERIAL PRIMARY KEY,
